chore: add tests for updates
